requirements.txt
alembic==1.4.2
aniso8601==8.1.0
appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
cachelib==0.1.1
certifi==2020.4.5.1
cffi==1.7.0
chardet==3.0.4
Click==7.0
crypto==1.4.1
cssselect==1.1.0
docopt==0.6.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Sitemap==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
howdoi==1.2.1
idna==2.9
importlib-metadata==2.0.0
IPy==1.1
itchat==1.3.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonify==0.5
lxml==4.5.0
Mako==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==1.4.6
Naked==0.1.31
openpyxl==3.0.4
peewee==3.13.3
pipreqs==0.4.10
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.6.1
PyMySQL==0.9.3
pypng==0.0.20
PyQRCode==1.2.1
pyquery==1.4.1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
qiniu==7.2.6
requests==2.23.0
shellescape==3.8.1
six==1.14.0
soupsieve==2.1
SQLAlchemy==1.3.16
tzlocal==2.1
urljoin==1.0.0
urllib3==1.25.9
Werkzeug==1.0.0
WTForms==2.2.1
wxpusher==2.2.0
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
yarg==0.1.9
zipp==3.4.0
